feat: add userIndicator and position status

+++ src/utils/index.ts
@@ -14,9 +14,9 @@ import { object, string } from "yup";
 export function prettify(value: string, n?: number) {
   const [front, back] = value.split(".");
   if (back) {
-    return parseInt(front).toLocaleString() + "." + back.slice(0, n ?? 2);
+    return parseFloat(front).toLocaleString() + "." + back.slice(0, n ?? 2);
   }
-  return parseInt(front).toLocaleString();
+  return parseFloat(front).toLocaleString();
 }
 
 export const lendingStatusRequestSchema = object({
